requirements.txt
pydantic-settings~=2.1
